// server/lib/ocr/klippa/types.ts
export type KlippaParseAccountingDocumentResponse = {
  request_id: string;
  result: string;
  data: {
    raw_text: string;
    fields: null;
    vision_files: null;
    hash: string;
    hash_duplicate: boolean;
    quality: {
      blurriness: number;
      over_exposed: number;
      under_exposed: number;
      characters: number;
      unsupported_chars: number;
      non_text_chars: number;
      unknown_document_type: number;
    };
    parsed: {
// ...
      document_type: string;
      full_document_type: Array<{
        classification: string;
        confidence: number;
      }>;
      hash: string;
      hash_duplicate: boolean;
      invoice_number: string;
      invoice_type: string;
      lines: Array<{
        description: string;
        lineitems: Array<{
          amount: number;
          amount_each: number;
          amount_ex_vat: number;
          description: string;
          discount_amount: number;
          discount_code: string;
          discount_percentage: number;
          ean: string;
          line_number: string;
          matched_purchase_order: {
            matched_amount: boolean;
            matched_quantity: boolean;
            matched_sku: boolean;
            matched_title: boolean;
            matched_title_confidence: number;
            purchase_order_identifier: string;
            purchase_order_line_item_identifier: string;
          };
          order_number: string;
          quantity: number;
          sku: string;
          title: string;
          unit_of_measurement: string;
          vat_amount: number;
          vat_code: string;
          vat_percentage: number;
        }>;
      }>;
// ...
      merchant_name: string;
// ...
    };
  };
};